Extract GeometryStatus component from DipoleControl.tsx into a separate file
